refactor CredentialParserContext
