+++ src/context/CredentialsContextProvider.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback, useContext } from 'react';
 import { getItem } from '../indexedDB';
 import SessionContext from './SessionContext';
 import { compareBy, reverse } from '../util';
-import CredentialParserContext from "./CredentialParserContext";
+import CredentialParserContext from './CredentialParserContext';
 import { initializeCredentialEngine } from "../lib/initializeCredentialEngine";
 import { CredentialVerificationError } from 'core/dist/error';
 import { useHttpProxy } from '../lib/services/HttpProxy/HttpProxy';
